More dds header refactor
